configure riverpod
