+++ lib/core/flavor/flavor.dart
@@ -1,0 +1,6 @@
+enum Flavor {
+  dev,
+  qa,
+  uat,
+  prod,
+}
