+++ lib/main_uat.dart
@@ -1,8 +1,8 @@
-import 'package:flutter/material.dart';
-import 'app.dart';
-import 'flavors.dart';
+
+import 'package:ecom_app/core/flavor/flavor.dart';
+import 'package:ecom_app/main.dart';
+
 
 void main() {
-  F.appFlavor = Flavor.UAT;
-  runApp(App());
+  mainApp(Flavor.uat);
 }
