--- lib/pages/my_home_page.dart
@@ -1,18 +1,0 @@
-import 'package:flutter/material.dart';
-import '../flavors.dart';
-
-class MyHomePage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(F.title),
-      ),
-      body: Center(
-        child: Text(
-          'Hello ${F.title}',
-        ),
-      ),
-    );
-  }
-}
